Move run commands to commands.ts
